+++ tests/unit/type/UnionTypeTest.php
@@ -30,27 +30,6 @@ final class UnionTypeTest extends TestCase
         $this->assertSame('bool|int', $type->name());
     }
 
-    public function testCanBeRepresentedAsStringForReturnTypeDeclaration(): void
-    {
-        $type = new UnionType(
-            Type::fromName('bool', false),
-            Type::fromName('int', false)
-        );
-
-        $this->assertSame(': bool|int', $type->getReturnTypeDeclaration());
-    }
-
-    public function testCanBeRepresentedAsStringForNullableReturnTypeDeclaration(): void
-    {
-        $type = new UnionType(
-            Type::fromName('bool', false),
-            Type::fromName('int', false),
-            Type::fromName('null', true)
-        );
-
-        $this->assertSame(': bool|int|null', $type->getReturnTypeDeclaration());
-    }
-
     public function testMayAllowNull(): void
     {
         $type = new UnionType(