+++ src/Type.php
@@ -17,6 +17,10 @@ abstract class Type
 {
     public static function fromValue($value, bool $allowsNull): self
     {
+        if ($value === false) {
+            return new FalseType;
+        }
+
         $typeName = gettype($value);
 
         if ($typeName === 'object') {
@@ -37,6 +41,9 @@ abstract class Type
         switch (strtolower($typeName)) {
             case 'callable':
                 return new CallableType($allowsNull);
+
+            case 'false':
+                return new FalseType;
 
             case 'iterable':
                 return new IterableType($allowsNull);